Consolidate integration tests into QueryTests.cs
