--- tests/ParadeDB.EntityFrameworkCore.IntegrationTests/Query/MatchAllTests.cs
@@ -1,180 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using ParadeDB.EntityFrameworkCore.Extensions;
-using Shouldly;
-
-namespace ParadeDB.EntityFrameworkCore.IntegrationTests.Query;
-
-public sealed class MatchAllTests : TestBase
-{
-    [Test]
-    public async Task MatchAll_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, "these"))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithInlineArray_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, new[] { "these", "shoes" }))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithArrayVariable_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        string[] terms = ["these", "shoes"];
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, terms))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithFuzzy_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Fuzzy("these", 2)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithInlineArrayAndFuzzy_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p =>
-                EF.Functions.MatchAll(p.Description, Pdb.Fuzzy(new[] { "these", "shoes" }, 2))
-            )
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithArrayVariableAndFuzzy_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        string[] terms = ["these", "shoes"];
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Fuzzy(terms, 2)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithBoost_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Boost("these", 2.3f)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithInlineArrayAndBoost_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p =>
-                EF.Functions.MatchAll(p.Description, Pdb.Boost(new[] { "these", "shoes" }, 2.3f))
-            )
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithArrayVariableAndBoost_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        string[] terms = ["these", "shoes"];
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Boost(terms, 2.3f)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithConst_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Const("these", 20.3f)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithInlineArrayAndConst_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p =>
-                EF.Functions.MatchAll(p.Description, Pdb.Const(new[] { "these", "shoes" }, 20.3f))
-            )
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithArrayVariableAndConst_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        string[] terms = ["these", "shoes"];
-
-        var results = await context
-            .Products.Where(p => EF.Functions.MatchAll(p.Description, Pdb.Const(terms, 20.3f)))
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-
-    [Test]
-    public async Task MatchAll_WithFuzzyAndBoost_ExecutesSuccessfully()
-    {
-        await using var context = DbFixture.CreateContext();
-
-        var results = await context
-            .Products.Where(p =>
-                EF.Functions.MatchAll(p.Description, Pdb.Boost(Pdb.Fuzzy("these", 2), 2.3f))
-            )
-            .ToListAsync();
-
-        results.ShouldNotBeNull();
-    }
-}
